Revert separation if game icons into left and right divs
I originally wanted each icon to be in a separate div so that if they changed sizes then they wouldn't shift around in the icons container. However, this caused problems with the JavaScript code responsible for appending the new icons to the HTML. When considering mobile responsiveness, I decided to remove the transformation effect that would scale the icons for win/loss. This looks a lot cleaner and means I don't have to modify the JavaScript.